requirements.txt

django
django-crispy-forms
crispy-bootstrap5
django-filter
djangorestframework
python-decouple
whitenoise
gunicorn
psycopg2-binary
openpyxl
reportlab